refactor: extract tool policy config helpers
